# deploy: Check signature statuses in parallel