#### Problem

During deployment, we need to check the status of sent transactions in order to see if they have been confirmed. If they are confirmed, they are removed from the resend list.

Unfortunately, the code currently checks signature statuses in series, which means it can take a long time to find out the status of thousands of transactions.

#### Summary of Changes

Check signature statuses concurrently. There isn't a huge impact in testing 1000 self-transfers on testnet, but here are some timings:

* after: 26s, 17s, 16s
* before: 25s, 15s, 20s

If you think this isn't worth it, I'm happy to close, but let me know!

<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->