@KirillLykov 2024-05-17
What if you try `JoinSet` instead? Wonder what would be the performance:

```
diff --git client/Cargo.toml client/Cargo.toml
index ece0b..b0652 100644
--- client/Cargo.toml
+++ client/Cargo.toml
@@ -35,6 +35,7 @@ solana-tpu-client = { workspace = true, features = ["default"] }
 solana-udp-client = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
+itertools = {workspace = true}
 
 [dev-dependencies]
 crossbeam-channel = { workspace = true }
diff --git client/src/send_and_confirm_transactions_in_parallel.rs client/src/send_and_confirm_transactions_in_parallel.rs
index aa65ef..35e73 100644
--- client/src/send_and_confirm_transactions_in_parallel.rs
+++ client/src/send_and_confirm_transactions_in_parallel.rs
@@ -6,6 +6,7 @@ use {
     bincode::serialize,
     dashmap::DashMap,
     futures_util::future::join_all,
+    itertools::Itertools,
     solana_quic_client::{QuicConfig, QuicConnectionManager, QuicPool},
     solana_rpc_client::spinner::{self, SendTransactionProgress},
     solana_rpc_client_api::{
@@ -16,7 +17,7 @@ use {
     solana_sdk::{
         hash::Hash,
         message::Message,
-        signature::{Signature, SignerError},
+        signature::{self, Signature, SignerError},
         signers::Signers,
         transaction::{Transaction, TransactionError},
     },
@@ -113,11 +114,12 @@ fn create_transaction_confirmation_task(
     tokio::spawn(async move {
         // check transactions that are not expired or have just expired between two checks
         let mut last_block_height = current_block_height.load(Ordering::Relaxed);
+        let mut futures = tokio::task::JoinSet::new();
 
         loop {
             if !unconfirmed_transaction_map.is_empty() {
                 let current_block_height = current_block_height.load(Ordering::Relaxed);
-                let transactions_to_verify: Vec<Signature> = unconfirmed_transaction_map
+                let transactions_to_verify: Vec<Vec<Signature>> = unconfirmed_transaction_map
                     .iter()
                     .filter(|x| {
                         let is_not_expired = current_block_height <= x.last_valid_block_height;
@@ -127,11 +129,23 @@ fn create_transaction_confirmation_task(
                         is_not_expired || is_recently_expired
                     })
                     .map(|x| *x.key())
+                    .chunks(MAX_GET_SIGNATURE_STATUSES_QUERY_ITEMS)
+                    .into_iter()
+                    .map(|chunk| chunk.collect::<Vec<_>>())
                     .collect();
-                for signatures in
-                    transactions_to_verify.chunks(MAX_GET_SIGNATURE_STATUSES_QUERY_ITEMS)
-                {
-                    if let Ok(result) = rpc_client.get_signature_statuses(signatures).await {
+                while futures.join_next().await.is_some() {}
+
+                for signatures in transactions_to_verify {
+                    let rpc_client = rpc_client.clone();
+                    let errors_map = errors_map.clone();
+                    let num_confirmed_transactions = num_confirmed_transactions.clone();
+                    let unconfirmed_transaction_map = unconfirmed_transaction_map.clone();
+                    //let signatures = signatures.to_vec();
+                    futures.spawn(async move {
+                        let Ok(result) = rpc_client.get_signature_statuses(&signatures).await
+                        else {
+                            return;
+                        };
                         let statuses = result.value;
                         for (signature, status) in signatures.iter().zip(statuses.into_iter()) {
                             if let Some((status, data)) = status
@@ -153,12 +167,12 @@ fn create_transaction_confirmation_task(
                                 }
                             };
                         }
-                    }
+                    });
                 }
 
                 last_block_height = current_block_height;
             }
-            tokio::time::sleep(Duration::from_secs(1)).await;
+            //tokio::time::sleep(Duration::from_secs(1)).await;
         }
     })
 }
 ```