@joncinque 2024-05-17
Sorry, I think `Task` is a little overkill in this case, since we don't need to detach the execution, we just need to do the network calls concurrently. Is there a reason that normal `Future` isn't good enough?