@KirillLykov 2024-05-27
> Sorry, I think `Task` is a little overkill in this case, since we don't need to detach the execution, we just need to do the network calls concurrently. Is there a reason that normal `Future` isn't good enough?

I think there is some misunderstanding of the patch intention. I was curious if there any performance boost due to `JoinSet` usage, if not than I would not go for it because it adds some complexity. 

Independently on this particular code, I think there are some benefits of using `JoinSet` vs `join_all!` when the workload is complex:
* spawning a task is cheap, if we don't want to consume all the threads from runtime thread pool we can control this by checking the JoinSet size,
* I feel like `JoinSet` has a very clean interface in comparison to FuturesOrdered/Unordered.
* `join_all` uses `FuturesOrdered` which has some a bit obscure problems (see https://github.com/rust-lang/futures-rs/issues/2387). My understanding is that in this particular code the aforementioned problem shouldn't happen. The only hypothetical scenario I see is when response processing takes some time due to waiting for lock inside DashMap for too long time.