@joncinque 2024-10-28
Closing since the new TPU client will be the best option :smile: